requirements.txt
python-telegram-bot==21.6
web3==6.11.1
aiohttp==3.10.11
sqlalchemy==2.0.23
python-dotenv==1.0.0
networkx==3.2.1
eth-account==0.9.0
eth-keyfile==0.6.1
cryptography==41.0.7
mnemonic==0.20
requests==2.31.0
websockets==12.0
psycopg2-binary==2.9.9
base58==2.1.1
solana==0.30.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
asyncpg==0.29.0
pandas==2.1.1
numpy==1.24.3